Remove wrong file and Reinit Slider
