+++ src/components/shared/Slider/index.tsx
@@ -1,11 +1,20 @@
-import * as React from 'react';
-
+import React from 'react';
 import styled from 'styled-components/native';
 
-export interface SliderProps {
+const Container = styled.View`
+  flex: 1;
+  background-color: transparent;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+`;
+
+interface Props {
+  children?: any;
 }
 
-function SearchInput(props: SliderProps): React.ReactElement {
-  return (
-  );
+function Shared(props: Props): React.ReactElement {
+  return <Container>{props.children}</Container>;
 }
+
+export default Shared;
